Extract coerce_closure_fn_ty function
